Move enum Key to own file
