+++ crates/egui/src/data/mod.rs
@@ -1,4 +1,7 @@
 //! All the data sent between egui and the backend
 
 pub mod input;
+mod key;
 pub mod output;
+
+pub use key::Key;
